% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/history.R
\name{add_history_entry}
\alias{add_history_entry}
\title{Add a history entry}
\usage{
add_history_entry(history, entry)
}
\arguments{
\item{history}{Current history list}

\item{entry}{New entry to add}
}
\value{
Updated history list (entry prepended)
}
\description{
Add a history entry
}
\keyword{internal}
