% tinyrox says don't edit this manually, but it can't stop you!
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{glinty input proxy.}

\item{output}{glinty output proxy.}

\item{session}{glinty session.}
}
\value{
NULL (side effects only).
}
\description{
Server logic for the Earshot app.
}
\keyword{internal}
