% tinyrox says don't edit this manually, but it can't stop you!
\name{delete_history_entry}
\alias{delete_history_entry}
\title{Delete a history entry}
\usage{
delete_history_entry(history, id)
}
\arguments{
\item{history}{Current history list}

\item{id}{ID of entry to delete}
}
\value{
Updated history list with entry removed
}
\description{
Delete a history entry
}
\keyword{internal}
