% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/history.R
\name{load_history}
\alias{load_history}
\title{Load history from disk}
\usage{
load_history()
}
\value{
List of history entries (newest first), or empty list
}
\description{
Load history from disk
}
\keyword{internal}
