% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/history.R
\name{save_history}
\alias{save_history}
\title{Save history to disk}
\usage{
save_history(history)
}
\arguments{
\item{history}{List of history entries}
}
\description{
Save history to disk
}
\keyword{internal}
