% tinyrox says don't edit this manually, but it can't stop you!
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
NULL (side effects only).
}
\description{
Server logic for the Earshot Shiny app.
}
\keyword{internal}
