% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
NULL (side effects only).
}
\description{
Server logic for the Earshot Shiny app.
}
\keyword{internal}
