% tinyrox says don't edit this manually, but it can't stop you!
\name{app_ui}
\alias{app_ui}
\title{App UI}
\usage{
app_ui()
}
\value{
A glinty UI tree.
}
\description{
Create the Earshot app user interface.
}
\details{
Assets are served from inst/app/www under /static/ by run_app().

}
\keyword{internal}
