% tinyrox says don't edit this manually, but it can't stop you!
\name{earshot_dir}
\alias{earshot_dir}
\title{History Persistence Functions}
\usage{
earshot_dir()
}
\value{
Path to ~/.earshot directory
}
\description{
Functions for saving and loading transcription history.
}
\keyword{internal}
\keyword{Get}
\keyword{the}
\keyword{earshot}
\keyword{data}
\keyword{directory}
\keyword{internal}
